fix failing tests and move utils to populateDb utils
