+++ discovery-provider/plugins/notifications/src/__tests__/mappings/repostOfRepost.test.ts
@@ -10,12 +10,13 @@ import {
   createTracks,
   insertFollows,
   createPlaylists,
+  resetTests,
+  setUpTestDbProcessor
 } from "../../utils/populateDB";
 
 import { AppEmailNotification } from "../../types/notifications";
 import { renderEmail } from "../../email/notifications/renderEmail";
 import { EntityType } from "../../email/notifications/types";
-import { resetTests, setUpTestDbProcessor } from "../utils";
 
 describe("Repost Of Repost Notification", () => {
   let processor: Processor;
